Add files via upload
The first iteration of this tool was built in Excel. I primarily used the tool to calculate the savings incurred by making regular additional payments to the principal (rolled into the monthly Payment). 
Functionally, it uses the *Actual Payment* to invert the annuity equation to solve for periods. This represents the shortened period in which a loan can be paid off in, given the additional payments to the principal. We can then use a simple interest formula to calculate the total cost of the loan and the duration of the financial responsibility. 

*Actual Payment* is defined as the Lender Min Payment and the additional payment to principal. 

I Hope to translate the principals of this tool to R or Python and make it a more interactive user friendly tool.